updated types to match new db

+++ src/models/game.rs
@@ -40,7 +40,7 @@ pub struct GameWithJoinedSummary {
     pub players_count: Option<i64>,
     pub winners_count: Option<i64>,
 
-    pub winning_points: Option<i64>,
+    pub winning_points: Option<i32>,
     pub total_drops: Option<i64>,
 
     pub is_winner: Option<bool>,
@@ -67,6 +67,6 @@ pub struct GameWithHostedSummary {
     pub players_count: Option<i64>,
     pub winners_count: Option<i64>,
 
-    pub winning_points: Option<i64>,
+    pub winning_points: Option<i32>,
     pub total_drops: Option<i64>,
 }
